chore: new package for event reporting logic
