+++ pkg/test/case.go
@@ -14,8 +14,6 @@ import (
 	petname "github.com/dustinkirkland/golang-petname"
 	"github.com/thoas/go-funk"
 	corev1 "k8s.io/api/core/v1"
-	eventsv1 "k8s.io/api/events/v1"
-	eventsbeta1 "k8s.io/api/events/v1beta1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -27,6 +25,7 @@ import (
 	"github.com/kudobuilder/kuttl/pkg/kubernetes"
 	"github.com/kudobuilder/kuttl/pkg/report"
 	testutils "github.com/kudobuilder/kuttl/pkg/test/utils"
+	eventutils "github.com/kudobuilder/kuttl/pkg/test/utils/events"
 )
 
 // testStepRegex contains one capturing group to determine the index of a step file.
@@ -144,175 +143,15 @@ func (t *Case) NamespaceExists(namespace string) (bool, error) {
 	return ns.Name == namespace, nil
 }
 
-// byFirstTimestamp sorts a slice of events by first timestamp, using their involvedObject's name as a tie breaker.
-type byFirstTimestamp []eventsbeta1.Event
-
-func (o byFirstTimestamp) Len() int      { return len(o) }
-func (o byFirstTimestamp) Swap(i, j int) { o[i], o[j] = o[j], o[i] }
-
-func (o byFirstTimestamp) Less(i, j int) bool {
-	if o[i].ObjectMeta.CreationTimestamp.Equal(&o[j].ObjectMeta.CreationTimestamp) {
-		return o[i].Name < o[j].Name
-	}
-	return o[i].ObjectMeta.CreationTimestamp.Before(&o[j].ObjectMeta.CreationTimestamp)
-}
-
-// byFirstTimestampV1 sorts a slice of eventsv1 by first timestamp, using their involvedObject's name as a tie breaker.
-type byFirstTimestampV1 []eventsv1.Event
-
-func (o byFirstTimestampV1) Len() int      { return len(o) }
-func (o byFirstTimestampV1) Swap(i, j int) { o[i], o[j] = o[j], o[i] }
-
-func (o byFirstTimestampV1) Less(i, j int) bool {
-	if o[i].ObjectMeta.CreationTimestamp.Equal(&o[j].ObjectMeta.CreationTimestamp) {
-		return o[i].Name < o[j].Name
-	}
-	return o[i].ObjectMeta.CreationTimestamp.Before(&o[j].ObjectMeta.CreationTimestamp)
-}
-
-// byFirstTimestampCoreV1 sorts a slice of corev1 by first timestamp, using their involvedObject's name as a tie breaker.
-type byFirstTimestampCoreV1 []corev1.Event
-
-func (o byFirstTimestampCoreV1) Len() int      { return len(o) }
-func (o byFirstTimestampCoreV1) Swap(i, j int) { o[i], o[j] = o[j], o[i] }
-
-func (o byFirstTimestampCoreV1) Less(i, j int) bool {
-	if o[i].ObjectMeta.CreationTimestamp.Equal(&o[j].ObjectMeta.CreationTimestamp) {
-		return o[i].Name < o[j].Name
-	}
-	return o[i].ObjectMeta.CreationTimestamp.Before(&o[j].ObjectMeta.CreationTimestamp)
-}
-
 // CollectEvents gathers all events from namespace and prints it out to log
 func (t *Case) CollectEvents(namespace string) {
+	ctx := context.TODO()
 	cl, err := t.Client(false)
 	if err != nil {
 		t.Logger.Log("Failed to collect events for %s in ns %s: %v", t.Name, namespace, err)
 		return
 	}
-
-	err = t.collectEventsV1(cl, namespace)
-	if err != nil {
-		t.Logger.Log("Trying with events eventsv1beta1 API...")
-		err = t.collectEventsBeta1(cl, namespace)
-		if err != nil {
-			t.Logger.Log("Trying with events corev1 API...")
-			err = t.collectEventsCoreV1(cl, namespace)
-			if err != nil {
-				t.Logger.Log("All event APIs failed")
-			}
-		}
-	}
-}
-
-func (t *Case) collectEventsBeta1(cl client.Client, namespace string) error {
-	eventsList := &eventsbeta1.EventList{}
-
-	err := cl.List(context.TODO(), eventsList, client.InNamespace(namespace))
-	if err != nil {
-		t.Logger.Logf("Failed to collect events for %s in ns %s: %v", t.Name, namespace, err)
-		return err
-	}
-
-	events := eventsList.Items
-	sort.Sort(byFirstTimestamp(events))
-
-	t.Logger.Logf("%s events from ns %s:", t.Name, namespace)
-	printEventsBeta1(events, t.Logger)
-	return nil
-}
-
-func (t *Case) collectEventsV1(cl client.Client, namespace string) error {
-	eventsList := &eventsv1.EventList{}
-
-	err := cl.List(context.TODO(), eventsList, client.InNamespace(namespace))
-	if err != nil {
-		t.Logger.Logf("Failed to collect events for %s in ns %s: %v", t.Name, namespace, err)
-		return err
-	}
-
-	events := eventsList.Items
-	sort.Sort(byFirstTimestampV1(events))
-
-	t.Logger.Logf("%s events from ns %s:", t.Name, namespace)
-	printEventsV1(events, t.Logger)
-	return nil
-}
-
-func (t *Case) collectEventsCoreV1(cl client.Client, namespace string) error {
-	eventsList := &corev1.EventList{}
-
-	err := cl.List(context.TODO(), eventsList, client.InNamespace(namespace))
-	if err != nil {
-		t.Logger.Logf("Failed to collect events for %s in ns %s: %v", t.Name, namespace, err)
-		return err
-	}
-
-	events := eventsList.Items
-	sort.Sort(byFirstTimestampCoreV1(events))
-
-	t.Logger.Logf("%s events from ns %s:", t.Name, namespace)
-	printEventsCoreV1(events, t.Logger)
-	return nil
-}
-
-func printEventsBeta1(events []eventsbeta1.Event, logger testutils.Logger) {
-	for _, e := range events {
-		// time type regarding action reason note reportingController related
-		logger.Logf("%s\t%s\t%s\t%s\t%s\t%s\t%s\t%s",
-			e.ObjectMeta.CreationTimestamp,
-			e.Type,
-			shortString(&e.Regarding),
-			e.Action,
-			e.Reason,
-			e.Note,
-			e.ReportingController,
-			shortString(e.Related))
-	}
-}
-
-func printEventsV1(events []eventsv1.Event, logger testutils.Logger) {
-	for _, e := range events {
-		// time type regarding action reason note reportingController related
-		logger.Logf("%s\t%s\t%s\t%s\t%s\t%s\t%s\t%s",
-			e.ObjectMeta.CreationTimestamp,
-			e.Type,
-			shortString(&e.Regarding),
-			e.Action,
-			e.Reason,
-			e.Note,
-			e.ReportingController,
-			shortString(e.Related))
-	}
-}
-
-func printEventsCoreV1(events []corev1.Event, logger testutils.Logger) {
-	for _, e := range events {
-		// time type regarding action reason note reportingController related
-		logger.Logf("%s\t%s\t%s\t%s\t%s\t%s\t%s\t%s",
-			e.ObjectMeta.CreationTimestamp,
-			e.Type,
-			shortString(&e.InvolvedObject),
-			e.Action,
-			e.Reason,
-			e.Message,
-			e.ReportingController,
-			shortString(e.Related))
-	}
-}
-
-func shortString(obj *corev1.ObjectReference) string {
-	if obj == nil {
-		return ""
-	}
-	fieldRef := ""
-	if obj.FieldPath != "" {
-		fieldRef = "." + obj.FieldPath
-	}
-	return fmt.Sprintf("%s %s%s",
-		obj.GroupVersionKind().GroupKind().String(),
-		obj.Name,
-		fieldRef)
+	eventutils.CollectAndLog(ctx, cl, namespace, t.Name, t.Logger)
 }
 
 // Run runs a test case including all of its steps.
